# internal/openapi/spec.yaml
openapi: "3.0.3"
info:
  title: SWAMP API
  description: Software Assurance Marketplace — AI-powered security analysis of Git repositories.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/v1/projects/{projectID}/packages/{packageID}:
    parameters:
      - $ref: "#/components/parameters/projectID"
      - $ref: "#/components/parameters/packageID"
    get:
      tags: [packages]
      summary: Get a package
      operationId: getPackage
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Package details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwarePackage"
    put:
      tags: [packages]
      summary: Update a package
      operationId: updatePackage
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageCreate"
      responses:
        "200":
          description: Package updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwarePackage"
    delete:
      tags: [packages]
      summary: Delete a package
      operationId: deletePackage
      security:
        - cookieAuth: []
      responses:
        "204":
          description: Package deleted

  # ── Analyses ──────────────────────────────────────────────
  /api/v1/projects/{projectID}/analyses:
    parameters:
      - $ref: "#/components/parameters/projectID"
    get:
      tags: [analyses]
      summary: List analyses for a project
      operationId: listAnalyses
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
    post:
      tags: [analyses]
      summary: Trigger a new analysis (write access)
      operationId: createAnalysis
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [package_ids]
              properties:
                package_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
      responses:
        "201":
          description: Analysis created and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
  /api/v1/projects/{projectID}/analyses/{analysisID}:
    parameters:
      - $ref: "#/components/parameters/projectID"
      - $ref: "#/components/parameters/analysisID"
    get:
      tags: [analyses]
      summary: Get analysis details
      operationId: getAnalysis
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/projects/{projectID}/analyses/{analysisID}/cancel:
    parameters:
      - $ref: "#/components/parameters/projectID"
      - $ref: "#/components/parameters/analysisID"
    post:
      tags: [analyses]
      summary: Cancel a running analysis
      operationId: cancelAnalysis
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Analysis cancelled
...
  /api/v1/api-keys:
    get:
      tags: [api-keys]
      summary: List your API keys
      operationId: listAPIKeys
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
    post:
      tags: [api-keys]
      summary: Create an API key
      operationId: createAPIKey
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                expires_in_days:
                  type: integer
                  minimum: 1
      responses:
        "201":
          description: API key created (secret is only shown once)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIKey"
                  - type: object
                    properties:
                      secret:
                        type: string
                        description: Full key (only returned at creation time)
  /api/v1/api-keys/{keyID}:
    parameters:
      - $ref: "#/components/parameters/keyID"
    delete:
      tags: [api-keys]
      summary: Revoke an API key
      operationId: revokeAPIKey
      security:
        - cookieAuth: []
      responses:
        "204":
          description: Key revoked

  # ── Admin ─────────────────────────────────────────────────
...
  /api/v1/admin/users/{userID}:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      tags: [admin]
      summary: Get a user
      operationId: getUser
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      tags: [admin]
      summary: Update a user
      operationId: updateUser
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                email:
                  type: string
                status:
                  type: string
                  enum: [active, suspended, disabled]
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags: [admin]
      summary: Delete a user
      operationId: deleteUser
      security:
        - cookieAuth: []
      responses:
        "204":
          description: User deleted
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: swamp_session
    bearerAuth:
      type: http
      scheme: bearer
      description: API key in Authorization header
  parameters:
...
    projectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    packageID:
      name: packageID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    analysisID:
      name: analysisID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    keyID:
      name: keyID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userID:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, suspended, disabled]
        last_login:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SoftwarePackage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        git_url:
          type: string
          format: uri
        git_branch:
          type: string
        git_commit:
          type: string
        analysis_prompt:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PackageCreate:
      type: object
      required: [name, git_url]
      properties:
        name:
          type: string
        git_url:
          type: string
          format: uri
        git_branch:
          type: string
          default: main
        analysis_prompt:
          type: string
    Analysis:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        triggered_by:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        status_detail:
          type: string
        agent_model:
          type: string
        environment:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    APIKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        revoked:
          type: boolean
...